+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.22.0 [#](https://github.com/idleberg/vscode-applescript/releases/tag/v0.22.0)
 
-- add option to terminate `osascript` / `osacompile` processes
+- add command to terminate `osascript` / `osacompile` processes
+- add command to open extension settings
 - add support for highlighter injection into Markdown code fences
 - update dependencies